+++ CHANGES.rst
@@ -3,7 +3,7 @@ Products.MeetingLalouviere Changelog
 
 The Products.MeetingCommunes version must be the same as the Products.PloneMeeting version
 
-4.1.5.4 (unreleased)
+4.1.5.4 (2021-04-12)
 --------------------
 
 - Use classifiers instead of categories for commissions.